+++ RealArtists.ShipHub.Common/GitHub/GitHubResponse.cs
@@ -41,7 +41,7 @@
     public GitHubResponse(GitHubRequest request) : base(request) {
     }
 
-    public ushort Pages { get; set; }
+    public uint Pages { get; set; }
 
     private bool _resultSet = false;
     private T _result = default(T);
